proiect final. repartizare componente
