+++ src/components/Loader.jsx
@@ -1,0 +1,3 @@
+export default function Loader() {
+  return <div className="spinner"></div>;
+}
