+++ src/components/Logo.jsx
@@ -1,0 +1,8 @@
+export default function Logo() {
+  return (
+    <div className="logo">
+      <span role="img">🍿</span>
+      <h1>FilmoSphere</h1>
+    </div>
+  );
+}
