+++ src/components/MainContent.jsx
@@ -1,0 +1,3 @@
+export default function MainContent({ children }) {
+  return <main className="main">{children}</main>;
+}
