+++ src/components/MovieList.jsx
@@ -1,0 +1,11 @@
+import Movie from "./Movie";
+
+export default function MovieList({ movies, onSelectMovie }) {
+  return (
+    <ul className="list list-movies">
+      {movies?.map((movie) => (
+        <Movie movie={movie} key={movie.imdbID} onSelectMovie={onSelectMovie} />
+      ))}
+    </ul>
+  );
+}
