+++ src/main.jsx
@@ -2,15 +2,9 @@ import React from "react";
 import ReactDOM from "react-dom/client";
 import App from "./App.jsx";
 import "./index.css";
-import StarRating from "./StartRating";
 
 ReactDOM.createRoot(document.getElementById("root")).render(
   <React.StrictMode>
     <App />
-    {/* <StarRating
-      maxRating={5}
-      messages={["Terrible", "Bad", "Okay", "Good", "Amazing!"]}
-    />
-    <StarRating size={2} color="red" className="test" /> */}
   </React.StrictMode>
 );
